Fix issue #189.
Previously the response headers and response were contained in a single
div - response-container. This was being emptied when the response was
being beautified, but only the response was added back, not the headers.
This commit adds new divs inside response-container so the code only
removes the response when beautifying, leaving the headers in place.